update on lesson 6
