--- lesson_6_dragla_marius/functions_wip.py
@@ -1,122 +1,0 @@
-from pprint import pprint
-
-description = ('Country', ['2011 ', '2012 ', '2013 ', '2014 ', '2015 ', '2016 ', '2017 ', '2018 ', '2019 '])
-
-raw_data = [
-    ('AL', [': ', ': ', ': ', ': ', ': ', ': ', ': ', '84 ', ': ']),
-    ('AT', ['75 ', '79 ', '81 ', '81 ', '82 ', '85 ', '89 ', '89 ', '90 ']),
-    ('BA', [': ', ': ', ': ', ': ', ': ', ': ', ': ', '69 ', '72 ']),
-    ('BE', ['77 ', '78 ', '80 ', '83 ', '82 ', '85 ', '86 ', '87 ', '90 ']),
-    ('BG', ['45 ', '51 ', '54 ', '57 ', '59 ', '64 ', '67 ', '72 ', '75 ']),
-    ('CH', [': ', ': ', ': ', '91 ', ': ', ': ', '93', ': ', '96 ']),
-    ('CY', ['57 ', '62 ', '65 ', '69 ', '71 ', '74 ', '79 ', '86 ', '90 ']),
-    ('CZ', ['67 ', '73 ', '73 ', '78 ', '79 ', '82 ', '83 ', '86 ', '87 ']),
-    ('DE', ['83 ', '85 ', '88 ', '89 ', '90 ', '92 ', '93 ', '94 ', '95 ']),
-    ('DK', ['90 ', '92 ', '93 ', '93 ', '92 ', '94 ', '97 ', '93 ', '95 ']),
-    ('EA', ['74 ', '76 ', '79 ', '81 ', '83 ', '85 ', '87 ', '89 ', '90 ']),
-    ('EE', ['69 ', '74 ', '79 ', '83 ', '88 ', '86 ', '88 ', '90 ', '90 ']),
-    ('EL', ['50 ', '54 ', '56 ', '66 ', '68 ', '69 ', '71 ', '76 ', '79 ']),
-    ('ES', ['63 ', '67 ', '70 ', '74 ', '79 ', '82 ', '83 ', '86 ', '91 ']),
-    ('FI', ['84 ', '87 ', '89 ', '90 ', '90 ', '92 ', '94 ', '94 ', '94 ']),
-    ('FR', ['76 ', '80 ', '82 ', '83 ', '83 ', '86 ', '86 ', '89 ', '90 ']),
-    ('HR', ['61 ', '66 ', '65 ', '68 ', '77 ', '77 ', '76 ', '82 ', '81 ']),
-    ('HU', ['63 ', '67 ', '70 ', '73 ', '76 ', '79 ', '82 ', '83 ', '86 ']),
-    ('IE', ['78 ', '81 ', '82 ', '82 ', '85 ', '87 ', '88 ', '89 ', '91 ']),
-    ('IS', ['93 ', '95 ', '96 ', '96 ', ': ', ': ', '98 ', '99 ', '98 ']),
-    ('IT', ['62 ', '63 ', '69 ', '73 ', '75 ', '79 ', '81 ', '84 ', '85 ']),
-    ('LT', ['60 ', '60 ', '65 ', '66 ', '68 ', '72 ', '75 ', '78 ', '82 ']),
-    ('LU', ['91 ', '93 ', '94 ', '96 ', '97 ', '97 ', '97 ', '93 b', '95 ']),
-    ('LV', ['64 ', '69 ', '72 ', '73 ', '76 ', '77 ', '79 ', '82 ', '85 ']),
-    ('ME', [': ', '55 ', ': ', ': ', ': ', ': ', '71 ', '72 ', '74 ']),
-    ('MK', [': ', '58 ', '65 ', '68 ', '69 ', '75 ', '74 ', '79 ', '82 ']),
-    ('MT', ['75 ', '77 ', '78 ', '80 ', '81 ', '81 ', '85 ', '84 ', '86 ']),
-    ('NL', ['94 ', '94 ', '95 ', '96 ', '96 ', '97 ', '98 ', '98 ', '98 ']),
-    ('NO', ['92 ', '93 ', '94 ', '93 ', '97 ', '97 ', '97 ', '96 ', '98 ']),
-    ('PL', ['67 ', '70 ', '72 ', '75 ', '76 ', '80 ', '82 ', '84 ', '87 ']),
-    ('PT', ['58 ', '61 ', '62 ', '65 ', '70 ', '74 ', '77 ', '79 ', '81 ']),
-    ('RO', ['47 ', '54 ', '58 ', '61 ', '68 ', '72 ', '76 ', '81 ', '84 ']),
-    ('RS', [': ', ': ', ': ', ': ', '64 ', ': ', '68 ', '73 ', '80 ']),
-    ('SE', ['91 ', '92 ', '93 ', '90 ', '91 ', '94', '95 ', '93 ', '96 ']),
-    ('SI', ['73 ', '74 ', '76 ', '77 ', '78 ', '78 ', '82 ', '87 ', '89 ']),
-    ('SK', ['71 ', '75 ', '78 ', '78 ', '79 ', '81 ', '81 ', '81 ', '82 ']),
-    ('TR', [': ', '47 ', '49 ', '60 ', '70 ', '76 ', '81 ', '84 ', '88 ']),
-    ('UK', ['83 ', '87 ', '88 ', '90 ', '91 ', '93 ', '94 ', '95 ', '96 ']),
-    ('XK', [': ', ': ', ': ', ': ', ': ', ': ', '89 ', '93 ', '93 ']),
-]
-
-_, year = description
-
-raw_data_unzip = ([i for i, j in raw_data],
-                 [j for i, j in raw_data])
-country, coverage = raw_data_unzip
-
-
-def dataset_func(raw_data):
-    dataset = {
-        country: [
-            {'year': year_val.strip(), 'coverage': coverage_val.strip()}
-            for coverage_val, year_val in zip(coverage, year)
-        ]
-        for country, coverage in raw_data
-    }
-
-    pprint(dataset)
-
-    return dataset
-
-
-# get_year_data(dataset, "2019")
-#
-# {'2019': [('Romania', 84), ('Germany', 95), ..., ('Latvia', 85)]}
-
-
-def get_year_data(dataset, year):
-    prep_year_data = {
-        year: [
-            [country, data_cov['coverage']]
-            for country in dataset
-            for data_cov in dataset[country]
-            if data_cov['year'] == year
-        ]
-    }
-    return prep_year_data
-
-
-# get_country_data(dataset, "Romania")
-# {'Romania': [('2019', 84), ('2018', 86), ..., ('2011', 72)]}
-# country_data = get_country_data(dataset, "Romania")
-
-def get_country_data(dataset, country):
-    prep_country_data = dataset.get(country)
-    return {
-        country: [
-            {value["year"], value["coverage"]}
-            for value in prep_country_data
-        ]
-    }
-
-
-# perform_average(country_data['Romania'])
-
-# still to be solved:
-
-# def perform_average(dataset):
-#     coverage_data = []
-#     coverage_sum = sum(coverage_data)
-#     coverage_len = len(coverage_data)
-#     average = coverage_sum / coverage_len
-#
-#     print(coverage_data)
-#
-#     return coverage_data
-
-
-dataset = dataset_func(raw_data)
-
-year_data = get_year_data(dataset, "2019")
-pprint(year_data)
-
-country_data = get_country_data(dataset, "RO")
-pprint(country_data)
-
-# perform_average(country_data['RO'])
